Move FieldPanel to wagtail.admin.panels.field_panel
